Refactor: Move DonutChart to shared folder
