+++ src/components/v5/frame/ColonyHome/partials/ReputationChart/ReputationChart.tsx
@@ -9,11 +9,11 @@ import { useColonyContext } from '~hooks';
 import { notNull } from '~utils/arrays';
 import { formatText } from '~utils/intl';
 import { ACTION_TYPE_FIELD_NAME } from '~v5/common/ActionSidebar/consts';
-import DonutChart from '~v5/common/DonutChart';
-import { ChartData } from '~v5/common/DonutChart/types';
 import TeamReputationSummaryRow from '~v5/common/TeamReputationSummary/partials/TeamReputationSummaryRow';
 import WidgetBox from '~v5/common/WidgetBox';
 import EmptyWidgetState from '~v5/common/WidgetBox/partials/EmptyWidgetState';
+import DonutChart from '~v5/shared/DonutChart';
+import { ChartData } from '~v5/shared/DonutChart/types';
 
 import { summaryLegendColor } from './consts';
 import { getTeamReputationChartData } from './utils';
